Update the Buffer constructor to avoid allocating unused memory
The Buffer constructor currently takes an allocated vector of the max length the buffer can support. This means that the memory consumption is the vector allocated memory even if its effectlve length (the buffer length) is much lower.

In order to avoid this extra memory usage, the constructor is changed to take the maximum length instead. The `cap` field of the buffer is also renamed to `max_length` which is more accurate because the capacity (aka allocated memory) might be larger due to the allocator doing over-provisioning.

The `resize` and `extend_from_slice` functions are updated accordingly. The `resize` is responsible for the memory allocation of the inner buffer; while it's not strictly necessary due to the actual implementation of `extend_from_slice` this makes the code more robust, and the calls to `resize` behaves as expected since it's part of the public API.