Rename the stream buffer constants and make the max buffer size configurable
The INITIAL_BUFFER_SIZE becomes BUFFER_CHUNK_SIZE which is more accurate as it's the max size that is use to read the data for filling the stream buffer.

BUFFER_SIZE becomes DEFAULT_MAX_BUFFER_SIZE as this is indeed the maximum size now. This value is now configurable by setting the `KARYON_MAX_BUFFER_SIZE` environment variable at compile time.

Since it uses a compilation value there can't be unit tests for the feature, however it has been tested manually by printing the output of `get_max_buffer_size()` after building with several values for the environment variable.